Refactor: Keep MyRadiobuttons separately
